Дорефачил контроллеры

+++ src/models/User.js
@@ -2,12 +2,12 @@ import Sequelize from "sequelize";
 import sequelize from "./Sequlize";
 
 const User = sequelize.define("user", {
-  UserId: {
-    type: Sequelize.FLOAT,
-    unique: true,
-  },
-  Login: Sequelize.STRING,
-  payExpiresDay: Sequelize.DATE
+    UserId: {
+        type: Sequelize.INTEGER,
+        unique: true,
+    },
+    Login: Sequelize.STRING,
+    payExpiresDay: Sequelize.DATE
 });
 
 export default User;
